+++ tools4j-collections/src/test/java/com/github/panpf/tools4j/collections/Arrayx2Test.kt
@@ -1,5 +1,0 @@
-package com.github.panpf.tools4j.collections
-
-class Arrayx2Test {
-    // todo Complete test
-}